[Chore] Refactor global models to resolve celery and import issues
- Move notifications package to services
- camelCase owner_photo logged activity field
