+++ src/api/endpoints/cohorts/crud.py
@@ -3,7 +3,7 @@ from flask import request
 
 from api.services.auth import token_required, roles_required
 from api.utils.helpers import response_builder
-from api.utils.marshmallow_schemas import base_schema
+from api.utils.marshmallow_schemas import basic_info_schema
 
 from .marshmallow_schemas import cohort_schema
 
@@ -12,7 +12,7 @@ class Cohorts(Resource):
     """Resource for cohorts crud operations."""
 
     def __init__(self, **kwargs):
-        """Inject dependacy for resource."""
+        """Inject dependencies for resource."""
         self.Cohort = kwargs['Cohort']
         self.Society = kwargs['Society']
 
@@ -38,14 +38,14 @@ class Cohorts(Resource):
         if not society:
             return response_builder(dict(
                 message="Error Invalid societyId."
-            ), 400)
+            ), 404)
 
         cohort = self.Cohort.query.filter_by(uuid=payload.get(
             'cohortId')).first()
         if not cohort:
             return response_builder(dict(
                 message="Error Invalid cohortId."
-            ), 400)
+            ), 404)
 
         if cohort.society_id == society.uuid:
             return response_builder(dict(
@@ -57,8 +57,8 @@ class Cohorts(Resource):
 
         cohort_data = cohort_schema.dump(cohort).data
         cohort_meta_data = {
-            'society': base_schema.dump(society).data,
-            'center': base_schema.dump(cohort.center).data
+            'society': basic_info_schema.dump(society).data,
+            'center': basic_info_schema.dump(cohort.center).data
         }
         cohort_data['meta'] = cohort_meta_data
