+++ src/api/endpoints/logged_activities/__init__.py
@@ -2,10 +2,8 @@
 
 
 def logged_activities_bp(Api, Blueprint):
-    from api.endpoints.activity_type.models import ActivityType
-    from api.endpoints.activities.models import Activity
+    from api.models import ActivityType, Activity, User
     from .models import LoggedActivity
-
     from .crud import LoggedActivitiesAPI
     from .user_logged_activities import UserLoggedActivitiesAPI
     from .approve import LoggedActivityApprovalAPI
@@ -37,7 +35,8 @@ def logged_activities_bp(Api, Blueprint):
         '/users/<string:user_id>/logged-activities/',
         endpoint='user_logged_activities',
         resource_class_kwargs={
-            'LoggedActivity': LoggedActivity
+            'LoggedActivity': LoggedActivity,
+            'User': User
         }
     )
