+++ src/api/endpoints/logged_activities/approve.py
@@ -13,7 +13,7 @@ class LoggedActivityApprovalAPI(Resource):
     decorators = [token_required]
 
     def __init__(self, **kwargs):
-        """Inject dependacy for resource."""
+        """Inject dependency for resource."""
         self.LoggedActivity = kwargs['LoggedActivity']
 
     @roles_required(["success ops"])
