+++ src/api/endpoints/logged_activities/reject.py
@@ -12,7 +12,7 @@ class LoggedActivityRejectionAPI(Resource):
     decorators = [token_required]
 
     def __init__(self, **kwargs):
-        """Inject dependacy for resource."""
+        """Inject dependency for resource."""
         self.LoggedActivity = kwargs['LoggedActivity']
 
     @roles_required(["success ops"])
