+++ src/api/endpoints/redemption_requests/models.py
@@ -1,4 +1,4 @@
-from api.models import Base
+from api.models.base import Base
 
 
 db = Base.db
