+++ src/api/endpoints/redemption_requests/redemption_numeration.py
@@ -3,7 +3,7 @@ from flask import request, g, current_app
 from flask_restful import Resource
 
 # imports from other packages
-from api.utils.notifications.email_notices import send_email
+from api.services.notifications.tasks import send_email
 from api.services.auth import token_required, roles_required
 from api.utils.helpers import response_builder
 from api.utils.marshmallow_schemas import basic_info_schema
@@ -24,7 +24,7 @@ class RedemptionRequestNumeration(Resource):
     """
 
     def __init__(self, **kwargs):
-        """Inject dependacy for resource."""
+        """Inject dependencies for resource."""
         self.RedemptionRequest = kwargs['RedemptionRequest']
         self.Society = kwargs['Society']
