+++ src/api/models/__init__.py
@@ -1,0 +1,10 @@
+from .base import Base
+from .center import Center
+from api.endpoints.activities.models import Activity
+from api.endpoints.activity_types.models import ActivityType
+from api.endpoints.cohorts.models import Cohort
+from api.endpoints.logged_activities.models import LoggedActivity
+from api.endpoints.redemption_requests.models import RedemptionRequest
+from api.endpoints.roles.models import Role
+from api.endpoints.societies.models import Society
+from api.endpoints.users.models import User
