+++ src/api/services/auth/auth.py
@@ -4,110 +4,15 @@ Authorisation Module.
 This module contains the authorisation required by the client to
 communicate with the API.
 """
-
 import base64
-import os
 from functools import wraps
-import requests
-from flask import current_app, g, jsonify, request, Response
+
+from flask import current_app, g, request
 from jose import ExpiredSignatureError, JWTError, jwt
 
-from api.endpoints.users.models import User
-from api.endpoints.roles.models import Role
-from api.endpoints.cohorts.models import Cohort
-from api.models import Center
+from .helpers import store_user_details
+from api.models import Role, User
 from api.utils.helpers import response_builder
-
-
-def auth_response(status_code, message):
-    """Build authentication responses."""
-    response = jsonify({
-        "message": message
-    })
-    response.status_code = status_code
-    return response
-
-
-def add_extra_user_info(
-    token,
-    user_id,
-    url=os.environ.get('ANDELA_API_URL')
-):  # pragma: no cover
-    """Retrive user information from ANDELA API.
-
-    params:
-        token(str): valid jwt token
-        user_id(str): id for user to retive information about
-
-    Return:
-        tuple(location, cohort, api_response)
-    """
-    cohort = location = None
-    Bearer = 'Bearer '
-    headers = {'Authorization': Bearer + token}
-
-    try:
-        api_response = requests.get(url + f"users/{user_id}", headers=headers)
-    except requests.exceptions.ConnectionError:
-        response = Response()
-        response.status_code = 503
-        response.json = lambda: {"Error": "Network Error."}
-        return cohort, location, response
-    except Exception:
-        response = Response()
-        response.status_code = 500
-        response.json = lambda: {"Error": "Something went wrong."}
-        return cohort, location, response
-
-    if api_response.status_code == 200 and api_response.json().get('cohort'):
-        cohort = Cohort.query.filter_by(
-            uuid=api_response.json().get('cohort').get('id')).first()
-        if not cohort:
-            cohort = Cohort(uuid=api_response.json().get('cohort').get('id'),
-                            name=api_response.json().get('cohort').get('name'))
-
-        location = Center.query.filter_by(
-            uuid=api_response.json().get('location').get('id')).first()
-        if not location:
-            location = Center(
-                uuid=api_response.json().get('location').get('id'))
-    return cohort, location, api_response
-
-
-def store_user_details(payload, token):
-    """Store user details in our database."""
-    user_id = payload["UserInfo"]["id"]
-    name = payload["UserInfo"]["name"]
-    email = payload["UserInfo"]["email"]
-    photo = payload["UserInfo"]["picture"]
-    roles = payload["UserInfo"]["roles"]
-
-    user = User.query.get(user_id)
-
-    # save user to db if they haven't been saved yet
-    if not user:
-        user = User(
-            uuid=user_id, name=name, email=email, photo=photo
-        )
-
-    cohort, location, _ = add_extra_user_info(token, user_id)
-
-    if cohort:
-        cohort.members.append(user)
-        cohort.save()
-        user.society_id = cohort.society.uuid if cohort.society else None
-
-    if location:
-        location.members.append(user)
-        location.save()
-
-    # set current user in flask global variable, g
-    user.roles = [
-        Role.query.filter_by(name=role.lower()).first() for role in roles
-        if role and role != "Andelan" and Role.query.filter_by(
-            name=role.lower()).first() is not None]
-    user.save()
-    return user
 
 
 def verify_token(authorization_token, public_key, audience=None, issuer=None):
@@ -221,4 +126,3 @@ def roles_required(roles):
             return f(*args, **kwargs)
         return decorated
     return check_user_role
-
