+++ src/manage.py
@@ -10,15 +10,10 @@ from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager, Shell, prompt_bool
 
 from api.utils.initial_data import generete_initial_data_run_time_env
-from api.models import db, Center
-from api.endpoints.cohorts.models import Cohort
-from api.endpoints.societies.models import Society
-from api.endpoints.activities.models import Activity
-from api.endpoints.roles.models import Role
-from api.endpoints.users.models import User
-
-from app import create_app
+from app import create_app, db
+from api.models import Center, Cohort, Society, Activity, Role, User
 from run_tests import test
+
 
 app = create_app(environment=os.environ.get('APP_SETTINGS', "Production"))
 manager = Manager(app)
